requirements.txt
Flask==1.1.2
pandas==1.2.3
pycryptodome==3.10.1
discord.py==2.1.0
dateparser==1.0.0
asyncio
gitpython
key